refactor(user/homepage): add everything user related inside grids/cards/crud sub-folders
