+++ FinEdgeFrontend/src/Components/HomePage/Grids/CategoriesGrid.tsx
