+++ FinEdgeFrontend/src/Components/HomePage/Grids/CategoriesInfo.tsx
