+++ FinEdgeFrontend/src/Components/HomePage/Grids/CategoriesPie.tsx
